Website used to search for and save movie titles and descriptions.
Movie Night is a website designed to allow users to search for movie
titles and information and then store those titles and information in a
phpMyAdmin database. When a user is ready to watch a movie they click
the "Let's Watch a Movie!" button and the website randomly selects a
movie from their database for the user to watch.